+++ script/core/signature.lua
@@ -81,6 +81,7 @@ local function makeSignatures(call, pos)
     local signs = {}
     local defs = vm.getDefs(node, 0)
     for _, src in ipairs(defs) do
+        src = guide.getObjectValue(src) or src
         if src.type == 'function'
         or src.type == 'doc.type.function' then
             signs[#signs+1] = makeOneSignature(src, oop, index)